requirements.txt
absl-py==2.2.0
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
etils==1.12.2
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
flax==0.10.4
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanize==4.12.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
libclang==18.1.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
narwhals==1.31.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.11.10
orjson==3.10.15
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
starlette==0.46.1
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.72
termcolor==2.5.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
treescope==0.1.9
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
